add test for lora
